Moved cryptographic operations
