+++ com.dynamo.cr/com.dynamo.cr.bob.test/src/com/dynamo/bob/bundle/test/BundlerTest.java
@@ -73,6 +73,7 @@ import com.dynamo.bob.archive.publisher.NullPublisher;
 import com.dynamo.bob.archive.publisher.PublisherSettings;
 import com.dynamo.bob.bundle.BundleHelper;
 import com.dynamo.bob.fs.DefaultFileSystem;
+import com.dynamo.bob.util.CryptographicOperations;
 import com.dynamo.liveupdate.proto.Manifest.HashAlgorithm;
 
 @RunWith(Parameterized.class)
@@ -534,8 +535,8 @@ public class BundlerTest {
     public void testAbsoluteCustomResourcePath() throws IOException, ConfigurationException, CompileExceptionError, MultipleCompileException, NoSuchAlgorithmException {
         final String expectedData = "dummy";
         final HashAlgorithm hashAlgo = HashAlgorithm.HASH_SHA1;
-        final byte[] expectedHash = ManifestBuilder.CryptographicOperations.hash(expectedData.getBytes(), hashAlgo);
-        final int hlen = ManifestBuilder.CryptographicOperations.getHashSize(hashAlgo);
+        final byte[] expectedHash = CryptographicOperations.hash(expectedData.getBytes(), hashAlgo);
+        final int hlen = CryptographicOperations.getHashSize(hashAlgo);
         int numBuiltins = createDefaultFiles(contentRoot);
         createFile(contentRoot, "game.project", "[project]\ncustom_resources=/m.txt\n[display]\nwidth=640\nheight=480\n");
         createFile(contentRoot, "m.txt", expectedData);
