+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/pipeline/GameProjectBuilder.java
@@ -488,7 +488,7 @@ public class GameProjectBuilder extends Builder<Void> {
             if (!privateKeyFileHandle.exists() || !publicKeyFileHandle.exists()) {
                 logger.info("No public or private key for manifest signing set in liveupdate settings or project options, generating keys instead.");
                 try {
-                    ManifestBuilder.CryptographicOperations.generateKeyPair(SignAlgorithm.SIGN_RSA, privateKeyFilepath, publicKeyFilepath);
+                    CryptographicOperations.generateKeyPair(SignAlgorithm.SIGN_RSA, privateKeyFilepath, publicKeyFilepath);
                 } catch (NoSuchAlgorithmException exception) {
                     throw new IOException("Unable to create manifest, cannot create asymmetric keypair!");
                 }
